Move "list_coordinators_of_cooperation" route into "user" folder
The list of coordinators page is a public information for all authenticated users and there are no differences between the views for different users, so there is no need to have the route implemented in different blueprints. The `UrlIndex.get_list_of_coordinators_url`-method has been simplified.
