+++ arbeitszeit_flask/accountant/routes.py
@@ -6,9 +6,6 @@ from flask import redirect, render_template
 from arbeitszeit.use_cases.approve_plan import ApprovePlanUseCase
 from arbeitszeit.use_cases.get_accountant_dashboard import GetAccountantDashboardUseCase
 from arbeitszeit.use_cases.get_plan_details import GetPlanDetailsUseCase
-from arbeitszeit.use_cases.list_coordinations_of_cooperation import (
-    ListCoordinationsOfCooperationUseCase,
-)
 from arbeitszeit.use_cases.list_plans_with_pending_review import (
     ListPlansWithPendingReviewUseCase,
 )
@@ -25,9 +22,6 @@ from arbeitszeit_web.www.presenters.get_accountant_dashboard_presenter import (
 )
 from arbeitszeit_web.www.presenters.get_plan_details_accountant_presenter import (
     GetPlanDetailsAccountantPresenter,
-)
-from arbeitszeit_web.www.presenters.list_coordinations_of_cooperation_presenter import (
-    ListCoordinationsOfCooperationPresenter,
 )
 from arbeitszeit_web.www.presenters.list_plans_with_pending_review_presenter import (
     ListPlansWithPendingReviewPresenter,
@@ -97,21 +91,3 @@ def plan_details(
         )
     else:
         return http_404()
-
-
-@AccountantRoute(
-    "/accountant/cooperation_summary/<uuid:coop_id>/coordinators", methods=["GET"]
-)
-def list_coordinators_of_cooperation(
-    coop_id: UUID,
-    list_coordinations_of_cooperation: ListCoordinationsOfCooperationUseCase,
-    presenter: ListCoordinationsOfCooperationPresenter,
-):
-    use_case_response = list_coordinations_of_cooperation.list_coordinations(
-        ListCoordinationsOfCooperationUseCase.Request(cooperation=coop_id)
-    )
-    view_model = presenter.list_coordinations_of_cooperation(use_case_response)
-    return render_template(
-        "accountant/list_coordinators_of_cooperation.html",
-        view_model=view_model,
-    )
