--- arbeitszeit_flask/templates/company/list_coordinators_of_cooperation.html
@@ -1,13 +1,0 @@
-{% extends "base.html" %}
-
-{% block navbar_start %}
-{% from 'macros/navbar.html' import navbar %}
-{{ navbar(view_model.navbar_items) }}
-{% endblock %}
-
-{% block content %}
-{% from 'macros/list_coordinations_of_cooperation.html' import list_coordinations_of_cooperation %}
-{{ list_coordinations_of_cooperation(view_model) }}
-{% endblock %}
-
-
