+++ arbeitszeit_web/url_index.py
@@ -34,9 +34,7 @@ class UrlIndex(Protocol):
     def get_coop_summary_url(self, *, coop_id: UUID) -> str:
         ...
 
-    def get_list_of_coordinators_url(
-        self, *, user_role: Optional[UserRole], cooperation_id: UUID
-    ) -> str:
+    def get_list_of_coordinators_url(self, *, cooperation_id: UUID) -> str:
         ...
 
     def get_company_dashboard_url(self) -> str:
